# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_monitoring-system_backend_internal_domain_inspection.BulkSaveResultRequest:
    properties:
      inspection_id:
        type: string
      results:
        items:
          $ref: '#/definitions/github_com_monitoring-system_backend_internal_domain_inspection.SaveResultRequest'
        type: array
    required:
    - inspection_id
    - results
    type: object
  github_com_monitoring-system_backend_internal_domain_inspection.CreateInspectionRequest:
    properties:
      area_id:
        type: string
      detail_kawasan_id:
        type: string
      kawasan_id:
        type: string
    required:
    - area_id
    - detail_kawasan_id
    - kawasan_id
    type: object
  github_com_monitoring-system_backend_internal_domain_inspection.InspectionStatus:
    enum:
    - Draft
    - Ongoing
    - Completed
    - Approved
    type: string
    x-enum-varnames:
    - InspectionStatusDraft
    - InspectionStatusOngoing
    - InspectionStatusCompleted
    - InspectionStatusApproved
  github_com_monitoring-system_backend_internal_domain_inspection.SaveResultRequest:
    properties:
      checking:
        enum:
        - OK
        - NG
        - NA
        type: string
      keterangan:
        type: string
      nilai:
        minimum: 0
        type: integer
      uraian_id:
        type: string
    required:
    - checking
    - uraian_id
    type: object
  github_com_monitoring-system_backend_internal_domain_inspection.UpdateInspectionStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/github_com_monitoring-system_backend_internal_domain_inspection.InspectionStatus'
        enum:
        - Draft
        - Ongoing
        - Completed
        - Approved
    required:
    - status
    type: object
...
  github_com_monitoring-system_backend_pkg_response.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: dev@company.com
    name: Internal Dev Team
  description: API untuk sistem monitoring audit internal perusahaan
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Monitoring Audit API
  version: "1.0"
paths:
...
  /inspection-results/{result_id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific inspection result
      parameters:
      - description: Result ID
        in: path
        name: result_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: result deleted
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete inspection result
      tags:
      - Inspection Results
    put:
      consumes:
      - application/json
      description: Update a specific inspection result
      parameters:
      - description: Result ID
        in: path
        name: result_id
        required: true
        type: string
      - description: Update result request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_monitoring-system_backend_internal_domain_inspection.SaveResultRequest'
      produces:
      - application/json
      responses:
        "200":
          description: result updated
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update inspection result
      tags:
      - Inspection Results
  /inspections:
    get:
      consumes:
      - application/json
      description: Get a paginated list of inspections with optional filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit per page
        in: query
        name: limit
        type: integer
      - description: Area ID filter
        in: query
        name: area_id
        type: string
      - description: Status filter
        in: query
        name: status
        type: string
      - description: Inspector ID filter
        in: query
        name: inspector_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: failed to fetch inspections
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all inspections
      tags:
      - Inspections
    post:
      consumes:
      - application/json
      description: Create a new inspection header
      parameters:
      - description: Inspection creation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_monitoring-system_backend_internal_domain_inspection.CreateInspectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: inspection created
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create inspection
      tags:
      - Inspections
  /inspections/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an inspection by ID
      parameters:
      - description: Inspection ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: inspection deleted
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete inspection
      tags:
      - Inspections
    get:
      consumes:
      - application/json
      description: Retrieve a specific inspection header by its ID
      parameters:
      - description: Inspection ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "404":
          description: inspection not found
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get inspection by ID
      tags:
      - Inspections
  /inspections/{id}/results:
    get:
      consumes:
      - application/json
      description: Retrieve all results associated with a specific inspection
      parameters:
      - description: Inspection ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: failed to fetch results
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get inspection results by inspection ID
      tags:
      - Inspection Results
  /inspections/{id}/results/bulk:
    post:
      consumes:
      - application/json
      description: Save multiple results for a specific inspection in bulk
      parameters:
      - description: Inspection ID
        in: path
        name: id
        required: true
        type: string
      - description: Bulk save request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_monitoring-system_backend_internal_domain_inspection.BulkSaveResultRequest'
      produces:
      - application/json
      responses:
        "200":
          description: results saved
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Bulk save inspection results
      tags:
      - Inspection Results
  /inspections/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of an existing inspection
      parameters:
      - description: Inspection ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_monitoring-system_backend_internal_domain_inspection.UpdateInspectionStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: status updated
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update inspection status
      tags:
      - Inspections
  /inspections/area/{areaId}/status:
    get:
      consumes:
      - application/json
      description: Calculate and get the inspection status progress for a specific
        area
      parameters:
      - description: Area ID
        in: path
        name: areaId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: failed to calculate area status
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get area status
      tags:
      - Inspections
  /issues/{id}/delegates:
    get:
      consumes:
      - application/json
      description: List all delegates for an issue
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Issue Delegates
      tags:
      - Issues
    post:
      consumes:
      - application/json
      description: Add a user as a delegate to an issue to allow them to follow up
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: string
      - description: Delegate Request
        in: body
        name: body
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Add Issue Delegate
      tags:
      - Issues
  /issues/{id}/delegates/{user_id}:
    delete:
      consumes:
      - application/json
      description: Remove a user from issue delegates
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: string
      - description: Delegate User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Remove Issue Delegate
      tags:
      - Issues
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"